#72 feat: MonthWidgetEntry 생성 및 MonthWidgetEntryView 수정
