+++ widget/Widget/MonthWidget.swift
@@ -1,0 +1,88 @@
+//
+//  MonthWidget.swift
+//  MonthWidget
+//
+//  Created by Kang Minsang on 2023/02/08.
+//  Copyright © 2023 FDEE. All rights reserved.
+//
+
+import WidgetKit
+import SwiftUI
+import Intents
+
+// MARK: Widget 내용 및 설정 부분
+struct MonthWidget: Widget {
+    let kind: String = "MonthWidget"
+
+    var body: some WidgetConfiguration {
+        IntentConfiguration(kind: kind, intent: ConfigurationIntent.self, provider: Provider()) { entry in
+            MonthWidgetEntryView(entry: entry)
+        }
+        .configurationDisplayName("Month widget".localized())
+        .description("Widget shows the top 5 tasks and the recording time by date.".localized())
+        .supportedFamilies([.systemMedium])
+    }
+}
+
+// MARK: Timeline 및 Entry 반환 부분
+struct Provider: IntentTimelineProvider {
+    func placeholder(in context: Context) -> MonthWidgetEntry {
+        MonthWidgetEntry(date: Date(), configuration: ConfigurationIntent())
+    }
+
+    func getSnapshot(for configuration: ConfigurationIntent, in context: Context, completion: @escaping (MonthWidgetEntry) -> ()) {
+        let entry = MonthWidgetEntry(date: Date(), configuration: configuration)
+        completion(entry)
+    }
+
+    func getTimeline(for configuration: ConfigurationIntent, in context: Context, completion: @escaping (Timeline<Entry>) -> ()) {
+        var entries: [MonthWidgetEntry] = []
+
+        // Generate a timeline consisting of five entries an hour apart, starting from the current date.
+        let currentDate = Date()
+        for hourOffset in 0 ..< 5 {
+            let entryDate = Calendar.current.date(byAdding: .hour, value: hourOffset, to: currentDate)!
+            let entry = MonthWidgetEntry(date: entryDate, configuration: configuration)
+            entries.append(entry)
+        }
+
+        let timeline = Timeline(entries: entries, policy: .atEnd)
+        completion(timeline)
+    }
+}
+
+// MARK: Widget 반환 부분
+struct MonthWidgetEntryView: View {
+    @Environment(\.widgetFamily) var family
+    var entry: Provider.Entry
+    
+    @ViewBuilder
+    var body: some View {
+        // MARK: 만약 동일한 위젯을 사이즈별로 제공하고 싶은 경우: switch family 로 분기처리 하면 된다.
+        MonthWidgetView()
+    }
+}
+
+struct MonthWidget_Previews: PreviewProvider {
+    static var previews: some View {
+        Group {
+            MonthWidgetEntryView(entry: MonthWidgetEntry(date: Date(), configuration: ConfigurationIntent()))
+                .previewDevice("iPad (9th generation)")
+                .previewContext(WidgetPreviewContext(family: .systemMedium))
+            
+            MonthWidgetEntryView(entry: MonthWidgetEntry(date: Date(), configuration: ConfigurationIntent()))
+                .previewDevice("iPhone 13 Pro")
+                .previewContext(WidgetPreviewContext(family: .systemMedium))
+            
+            MonthWidgetEntryView(entry: MonthWidgetEntry(date: Date(), configuration: ConfigurationIntent()))
+                .redacted(reason: .placeholder)
+                .previewDevice("iPhone 13 Pro")
+                .previewContext(WidgetPreviewContext(family: .systemMedium))
+            
+            MonthWidgetEntryView(entry: MonthWidgetEntry(date: Date(), configuration: ConfigurationIntent()))
+                .previewDevice("iPad Pro (12.9-inch) (6th generation)")
+                .previewContext(WidgetPreviewContext(family: .systemMedium))
+        }
+    }
+}
+
