+++ widget/Widget/MonthWidgetEntry.swift
@@ -1,0 +1,15 @@
+//
+//  MonthWidgetEntry.swift
+//  Project_Timer
+//
+//  Created by Kang Minsang on 2023/05/11.
+//  Copyright © 2023 FDEE. All rights reserved.
+//
+
+import Foundation
+import WidgetKit
+
+struct MonthWidgetEntry: TimelineEntry {
+    let date: Date
+    let configuration: ConfigurationIntent
+}
