+++ widget/en.lproj/Localizable.strings
@@ -7,3 +7,4 @@
 */
 
 "Month widget" = "Month widget";
+"Widget shows the top 5 tasks and the recording time by date." = "Widget shows the top 5 tasks and the recording time by date.";
