+++ widget/ko.lproj/Localizable.strings
@@ -7,3 +7,4 @@
 */
 
 "Month widget" = "Month 위젯";
+"Widget shows the top 5 tasks and the recording time by date." = "상위 5가지 Task와 날짜별 기록시간을 보여줍니다.";
